requirements_test.txt
boto3==1.23.7
PyHamcrest==2.0.3
vcrpy==4.1.1
